backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt<5.0
python-jose[cryptography]==3.3.0
google-auth==2.25.2
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0
clicksend-client==5.0.78
email-validator==2.1.0
pydantic[email]==2.5.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
